# packages/server/src/auth/dtos/login-google.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNotEmpty } from 'class-validator';

export class LoginGoogleDto {
  @ApiProperty({ description: 'Googleトークン' })
  @IsNotEmpty()
  token: string;
}