# packages/swagger/swagger.yml
openapi: 3.0.0
paths:
  /api:
    get:
      operationId: AppController_getHello
      parameters: []
      responses:
        '200':
          description: ''
  /healthz:
    get:
      operationId: HealthzController_getHealthz
      parameters: []
      responses:
        '200':
          description: ''
  /api/login/google:
    post:
      operationId: AuthController_loginGoogle
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginGoogleDto'
      responses:
        '200':
          description: ''
  /api/register/google:
    post:
      operationId: AuthController_registerGoogle
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginGoogleDto'
      responses:
        '201':
          description: ''
  /api/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      responses:
        '200':
          description: ''
  /api/books:
    get:
      operationId: BooksController_getBookList
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    post:
      operationId: BooksController_createBook
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
  /api/books/{id}:
    get:
      operationId: BooksController_getBook
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
    delete:
      operationId: BooksController_deleteBook
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: ''
  /api/books/self:
    post:
      operationId: BooksController_createBookSelf
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookSelfDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
  /api/books/memo/{id}:
    patch:
      operationId: BooksController_patchBookMemo
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchBookMemoDto'
      responses:
        '200':
          description: ''
  /api/books/status/{id}:
    patch:
      operationId: BooksController_patchBookStatus
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchBookStatusDto'
      responses:
        '200':
          description: ''
info:
  title: Reading Management API
  description: API Docs
  version: '1.0'
  contact: {}
tags: []
servers:
  - url: '{environment}'
    variables:
      environment:
        enum:
          - http://localhost:3000
          - https://api.bookshelf.yu-ta-9.com
        default: http://localhost:3000
components:
  schemas:
    LoginGoogleDto:
      type: object
      properties:
        token:
          type: string
          description: Googleトークン
      required:
        - token
    Book:
      type: object
      properties:
        id:
          type: number
        isbn:
          type: string
          description: ISBNコード
        name:
          type: string
          description: 書籍名
        author:
          type: string
          description: 著者名
        publisher:
          type: string
          description: 出版社名
        status:
          type: number
          description: '読書状態 1: 未読 2: 読書中 3: 読了 4: 中断'
          enum:
            - 1
            - 2
            - 3
            - 4
        category:
          type: number
          description: カテゴリー(0 = "未設定")
        image_path:
          type: string
          description: 画像パス
        memo:
          type: string
          description: メモ
        createdAt:
          format: date-time
          type: string
          description: 追加日
        updatedAt:
          format: date-time
          type: string
          description: 更新日
      required:
        - id
        - isbn
        - name
        - author
        - publisher
        - status
        - category
        - image_path
        - memo
        - createdAt
        - updatedAt
    CreateBookSelfDto:
      type: object
      properties:
        name:
          type: string
          description: 書籍名
        author:
          type: string
          description: 著者名
        publisher:
          type: string
          description: 出版社名
      required:
        - name
        - author
        - publisher
    CreateBookDto:
      type: object
      properties:
        isbn:
          type: string
          description: ISBNコード
      required:
        - isbn
    PatchBookMemoDto:
      type: object
      properties:
        memo:
          type: string
          description: メモ
      required:
        - memo
    PatchBookStatusDto:
      type: object
      properties:
        status:
          type: number
          description: 読書状態
      required:
        - status